@@ -1,34 +1,34 @@
 {
-  "name": "backend",
-  "private": true,
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite --port=3000",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@ckeditor/ckeditor5-build-classic": "^39.0.1",
-    "@ckeditor/ckeditor5-vue": "^5.1.0",
-    "axios": "^0.27.2",
-    "chart.js": "^3.9.1",
-    "sortablejs": "^1.15.0",
-    "sortablejs-vue3": "^1.2.9",
-    "uuid": "^9.0.0",
-    "vue": "^3.2.25",
-    "vue-chartjs": "^4.1.2",
-    "vue-router": "^4.0.13",
-    "vue3-treeselect": "^0.1.10",
-    "vuex": "^4.0.2"
-  },
-  "devDependencies": {
-    "@headlessui/vue": "^1.6.6",
-    "@heroicons/vue": "^1.0.6",
-    "@tailwindcss/forms": "^0.5.2",
-    "@vitejs/plugin-vue": "^2.3.3",
-    "autoprefixer": "^10.4.7",
-    "postcss": "^8.4.14",
-    "tailwindcss": "^3.1.5",
-    "vite": "^4.4.9"
-  }
+    "name": "backend",
+    "private": true,
+    "version": "0.0.0",
+    "scripts": {
+        "dev": "vite --port=3000",
+        "build": "vite build",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "@ckeditor/ckeditor5-build-classic": "^39.0.1",
+        "@ckeditor/ckeditor5-vue": "^5.1.0",
+        "axios": "^1.6.8",
+        "chart.js": "^3.9.1",
+        "sortablejs": "^1.15.2",
+        "sortablejs-vue3": "^1.2.11",
+        "uuid": "^9.0.1",
+        "vue": "^3.4.21",
+        "vue-chartjs": "^4.1.2",
+        "vue-router": "^4.3.0",
+        "vue3-treeselect": "^0.1.10",
+        "vuex": "^4.1.0"
+    },
+    "devDependencies": {
+        "@headlessui/vue": "^1.7.19",
+        "@heroicons/vue": "^1.0.6",
+        "@tailwindcss/forms": "^0.5.7",
+        "@vitejs/plugin-vue": "^5.0.4",
+        "autoprefixer": "^10.4.19",
+        "postcss": "^8.4.38",
+        "tailwindcss": "^3.4.1",
+        "vite": "^5.2.4"
+    }
 }
